`Refactor` Create @repo/db package
- With the need to share db connection between the main API and the Admin dashboard, I moved everything related to drizzle, db schemas and connections to a standalone `@repo/db` package
