+++ packages/db/drizzle.config.ts
@@ -1,8 +1,8 @@
 import type { Config } from "drizzle-kit";
-import { env } from "./src/env";
+import { env } from "./env";
 
 export default {
-    schema: "./src/db/schema/index.ts",
+    schema: "./src/schema/index.ts",
     dialect: "mysql",
     out: "./drizzle",
     dbCredentials: {
