+++ packages/db/package.json
@@ -1,0 +1,44 @@
+{
+    "name": "@repo/db",
+    "version": "0.0.0",
+    "type": "module",
+    "private": true,
+    "scripts": {
+        "dev": "tsc --watch",
+        "build": "tsc",
+        "migrate": "dotenv -- tsx src/migrate.ts",
+        "generate": "dotenv -- drizzle-kit generate",
+        "studio": "dotenv -- drizzle-kit studio"
+    },
+    "exports": {
+        "./app": {
+            "types": "./src/app.ts",
+            "default": "./dist/app.js"
+        },
+        "./cookies": {
+            "types": "./src/cookies.ts",
+            "default": "./dist/cookies.js"
+        },
+        "./connection": {
+            "types": "./src/connection.ts",
+            "default": "./dist/src/connection.js"
+        },
+        "./schema": {
+            "types": "./src/schema/index.ts",
+            "default": "./dist/src/schema/index.js"
+        }
+    },
+    "devDependencies": {
+        "@repo/typescript-config": "*",
+        "@types/eslint": "^8.56.5",
+        "@types/node": "^20.11.24",
+        "typescript": "^5.3.3"
+    },
+    "dependencies": {
+        "dotenv": "^16.4.5",
+        "drizzle-orm": "^0.38.3",
+        "drizzle-zod": "^0.7.0",
+        "zod": "^3.23.8",
+        "mysql2": "^3.13.0"
+    }
+}
