+++ packages/db/src/migrate.ts
@@ -1,11 +1,9 @@
-import postgres from "postgres";
 import chalk from "chalk";
 import path from "path";
 
 import { drizzle } from "drizzle-orm/mysql2";
 import { migrate } from "drizzle-orm/mysql2/migrator";
 import mysql from "mysql2/promise";
-
 import { env } from "../env";
 
 const dbUrl = env.DB_URL.replace("postgres:5432", "localhost:5432");
